[Changed] Behaviour to not overwrite source files per default
Per default the extension no longer overwrites source files, but still
gives the possibility for an opt-in for this behaviour by setting
tinypng.forceOverwrite setting to true.

Per default compressed images will be named like <sourcename>.min.<ext>, but
also this postfix can be changed by the user in the settings.